+++ packages/react-form-state/CHANGELOG.md
@@ -13,3 +13,4 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/).
 ### Fixed
 
 - Fixed Nested and List component race condition. Nested and List now pass a function to their `onChange` prop instead of an object so that the data object will be created within `setState`.
+- No longer accidentally import all of lodash.